requirements/requirements.txt
numpy
opencv-contrib-python
numba
scipy
matplotlib
tensorflow
torchfile
scikit-image
seaborn
torch
torchvision
fuzzy-c-means
oct2py
tensorflow-probability
tf_slim
pyntcloud
scikit-learn
open3d
imblearn
gdown
zipfile36 
h5py
tqdm
PilloW
torch-optimizer
retry
vector-quantize-pytorch
linear_attention_transformer
cmake
faiss-gpu
phasepack
pyfftw
lpips
torchmetrics
libsvm
Cython
pyiqa
dlib
pysaliency
pyhull
pyamg
brisque